requirements.txt
loguru==0.7.2
huggingface-hub==0.25.2
openai==1.35.13
Pillow==11.0.0
